+++ sdk/crypto/basic/crypt/openssl_crypt_cipher.cpp
@@ -172,11 +172,13 @@ return_t openssl_crypt::open(crypt_context_t **handle, crypt_algorithm_t algorit
         EVP_CIPHER_CTX_set_padding(context->encrypt_context, 1);
         EVP_CIPHER_CTX_set_padding(context->decrypt_context, 1);
 
+        // TLS
         if (ccm8 == mode) {
-            context->tsize = 8;
+            context->lsize = 3;  // CCM_SET_L 3, CCM_SET_IVLEN=15-L=12
+            context->tsize = 8;  // AEAD_SET_TAG 8
         } else if (ccm16 == mode) {
-            // TLS
-            context->tsize = 16;
+            context->lsize = 3;   // CCM_SET_L 3, CCM_SET_IVLEN=15-L=12
+            context->tsize = 16;  // AEAD_SET_TAG 16
         }
 
         *handle = context;
@@ -365,7 +367,7 @@ return_t openssl_crypt::encrypt_internal(crypt_context_t *handle, const unsigned
                  * the size of the authentication tag is fixed at 128 bits
                  */
                 tag_size = context->tsize ? context->tsize : 16;
-            } else if ((crypt_mode_t::ccm == context->mode) || (crypt_mode_t::ccm8 == context->mode)) {
+            } else if ((crypt_mode_t::ccm == context->mode) || (crypt_mode_t::ccm8 == context->mode) || (crypt_mode_t::ccm16 == context->mode)) {
                 tag_size = context->tsize ? context->tsize : 14;
                 uint16 lsize = context->lsize ? context->lsize : 8;
                 uint16 nonce_size = 15 - lsize;
@@ -374,21 +376,6 @@ return_t openssl_crypt::encrypt_internal(crypt_context_t *handle, const unsigned
                 // EVP_CTRL_CCM_SET_IVLEN for Nonce (15-L)
                 EVP_CIPHER_CTX_ctrl(context->encrypt_context, EVP_CTRL_CCM_SET_IVLEN, nonce_size, nullptr);
                 EVP_CIPHER_CTX_ctrl(context->encrypt_context, EVP_CTRL_AEAD_SET_TAG, tag_size, nullptr);
-
-                binary_t &key = context->datamap[crypt_item_t::item_cek];
-                EVP_CipherInit_ex(context->encrypt_context, nullptr, nullptr, &key[0], &iv[0], 1);
-
-                ret_cipher = EVP_CipherUpdate(context->encrypt_context, nullptr, &size_update, nullptr, plainsize);
-                if (1 > ret_cipher) {
-                    ret = errorcode_t::error_cipher;
-                    __leave2_trace_openssl(ret);
-                }
-            } else if (crypt_mode_t::ccm16 == context->mode) {
-                // TLS
-                tag_size = 16;  // resize
-
-                EVP_CIPHER_CTX_ctrl(context->encrypt_context, EVP_CTRL_CCM_SET_IVLEN, 12, nullptr);
-                EVP_CIPHER_CTX_ctrl(context->encrypt_context, EVP_CTRL_AEAD_SET_TAG, 16, nullptr);
 
                 binary_t &key = context->datamap[crypt_item_t::item_cek];
                 EVP_CipherInit_ex(context->encrypt_context, nullptr, nullptr, &key[0], &iv[0], 1);
@@ -541,22 +528,13 @@ return_t openssl_crypt::decrypt_internal(crypt_context_t *handle, const unsigned
                 __leave2_trace(ret);
             }
 
-            if ((crypt_mode_t::ccm) == context->mode || (crypt_mode_t::ccm8 == context->mode)) {
+            if ((crypt_mode_t::ccm) == context->mode || (crypt_mode_t::ccm8 == context->mode) || (crypt_mode_t::ccm16 == context->mode)) {
                 uint16 lsize = context->lsize ? context->lsize : 8;
                 uint16 nonce_size = 15 - lsize;
 
                 EVP_CIPHER_CTX_ctrl(context->decrypt_context, EVP_CTRL_CCM_SET_L, lsize, nullptr);
                 // EVP_CTRL_CCM_SET_IVLEN for Nonce (15-L)
                 EVP_CIPHER_CTX_ctrl(context->decrypt_context, EVP_CTRL_CCM_SET_IVLEN, nonce_size, nullptr);
-                EVP_CIPHER_CTX_ctrl(context->decrypt_context, EVP_CTRL_AEAD_SET_TAG, (*tag).size(), (void *)&(*tag)[0]);
-
-                binary_t &key = context->datamap[crypt_item_t::item_cek];
-                EVP_CipherInit_ex(context->decrypt_context, nullptr, nullptr, &key[0], &iv[0], 0);
-
-                ret_cipher = EVP_CipherUpdate(context->decrypt_context, nullptr, &size_update, nullptr, ciphersize);
-            } else if (crypt_mode_t::ccm16 == context->mode) {
-                // TLS
-                EVP_CIPHER_CTX_ctrl(context->decrypt_context, EVP_CTRL_CCM_SET_IVLEN, 12, nullptr);
                 EVP_CIPHER_CTX_ctrl(context->decrypt_context, EVP_CTRL_AEAD_SET_TAG, (*tag).size(), (void *)&(*tag)[0]);
 
                 binary_t &key = context->datamap[crypt_item_t::item_cek];