+++ test/crypto/sample.cpp
@@ -67,6 +67,7 @@ int main(int argc, char** argv) {
         test_crypto_aead();
         test_cbc_hmac_tls_mte();  // mac_then_encrypt
         test_cbc_hmac_tls_etm();  // encrypt_then_mac
+        test_aead_ccm();
     }
     __finally2 { openssl_cleanup(); }